day2 rough
even though I'm not racing for a leaderboard spot, I'm still timing
myself.

5.58 for part1, and then another 3.28 for part2

that's not fast enough for the leaderboard, but that's alright. I blame
this on Rust not being optimized for this sort of ad-hoc string munging.
Something like scanf would have made it a lot faster. ah well.