day4 rough
6:35 for part 1, and another 11:16 for part 2.

not a difficult question by any means - it's just a bunch of data
munging and validation. Rust is pretty good at this sort of stuff, but
the achilles heel is that there's far too much "line noise" required to
do certain things. e.g: defining a hashmap, specifying the types of
constant expressions (&[&str] is a big yikes), etc...

I'm sure the code will look wonderfully elegant after a round of
cleanup, but no, Rust is certainly not the language to use when racing
for a leaderboard spot on these sorts of input manipulation questions.