day5
weekend = lazy = no racing, even against myself

I was about to do the naiive thing which is to literally perform BSP on
0..128 and 0..8, but then I saw the trick (i.e: that the boarding passes
are just binary numbers, albeit in a somewhat unconventional encoding).
At which point, the solution was pretty trivial.

Part 2 did give me a moment of thought, though that was mostly because
I didn't really understand what the question was on my first reading.

I'm not _super_ happy with my solution for part 2, so I think I might
hack on it a bit more...