day5 using binary search
yeah, I knew there was a nicer way to do it.

I guess the whole "BSP" part of the question was _really_ a hint for
part 2, which is nearly optimized using a binary search.

Interestingly enough, the Rust standard library includes a (currently
nightly-only) `core::slice::partition_point` method which can be used to
find a partition point in a sorted array given a certain predicate.
Since the array is sorted, it is able to perform an efficient binary
search to find the partition point.

Unfortunately, this method can't be used directly, as the predicate
method only accepts the element to be checked, and not it's index (for
additional context). As such, I've added a custom SliceExt trait to my
prelude which implements a modified `partition_point_enumerated` method
which exposes the index alongside the element in the predicate function.

The end result is a very neat looking part 2 solution.