day6
weekend = lazy = no racing. same as yesterday.

yeah, this one was pretty decent as well. reminds me of an old adage I
lived by back when I was interviewing a bunch: when in doubt, use a
hashmap!

That said, since we know that the input is only comprised of the letters
a-z, there's a more efficient zero-allocation solution by using a
26-entry byte-array (or for part1, a 26-bit integer bit-array). But I've
got to go do something else now, so I that isn't something I'll be
looking in to.