day23 part1
yeah, brute force would certainly work to get an answer - back of the
envelope calculations put the `cups` array at ~64Mb, and modern
computers are reasonably fast. it'll take a bloody long time to get an
answer, but I _would_ get an answer.

That's pretty boring through, so I'll just shelve the problem for now
until I have some more time to work on it.